\name{CLp10c15}
\docType{data}
\alias{CLp10c15}
\title{Control Limits for data vector dimension p=3, quarantine period c
  = 15}
\description{
  This data set contains control limits for data vector dimension p = 10, quarantine period c
  = 15 for in control average run length (ARL) values: 100, 200, 500,
  1000, 2000.
}
\usage{CLp10c15}
\keyword{datasets}
