\name{CLp2c9}
\docType{data}
\alias{CLp2c9}
\title{Control Limits for data vector dimension p=2, quarantine period c
  = 2}
\description{
  This data set contains control limits for data vector dimension p = 2, quarantine period c
  = 9 for in control average run length (ARL) values: 100, 200, 500,
  1000, 2000.
}
\usage{CLp2c9}
\keyword{datasets}
