\name{CLp8c15}
\docType{data}
\alias{CLp8c15}
\title{Control Limits for data vector dimension p=3, quarantine period c
  = 15}
\description{
  This data set contains control limits for data vector dimension p = 8, quarantine period c
  = 15 for in control average run length (ARL) values: 100, 200, 500,
  1000, 2000.
}
\usage{CLp8c15}
\keyword{datasets}
