\name{Figure1tau150}
\docType{data}
\alias{Figure1tau150}
\title{Out of Control Average Run Length for shift location tau = 150}
\description{
  This data set contains out of control average run length (ARL) values
  for nonparametric multivariate change point model.  Shift location tau
  = 150 for shift size delta and several quarantine values (from Figure
  1).  
}
\usage{Figure1tau150}
\keyword{datasets}
