\name{Figure2Parametric}
\docType{data}
\alias{Figure2Parametric}
\title{Out of Control Average Run Length for Parametric Multivariate
  Change Point Model}
\description{
  This data set contains out of control average run length (ARL) values
  for parametric multivariate change point model (from Figure 2).  
}
\usage{Figure2Parametric}
\keyword{datasets}
